Refactor Plex SSO flow into dedicated handler
