+++ Cantinarr/Features/OverseerrUsers/Logic/OverseerrUsersViewModel.swift
@@ -106,7 +106,7 @@ class OverseerrUsersViewModel: ObservableObject {
     var service: OverseerrUsersService
     private let settingsKey: String
     private let authContext = OverseerrAuthContextProvider()
-    private var tokenKey: String { "plexAuthToken-\(settingsKey)" }
+    private let plexSSOHandler: PlexSSOHandler
     private var authCancellable: AnyCancellable?
     private var plexSSOTask: Task<Void, Never>?
     let keywordActivatedSubject = PassthroughSubject<Void, Never>()
@@ -120,9 +120,14 @@ class OverseerrUsersViewModel: ObservableObject {
         // Added callback
         self.service = service
         self.settingsKey = settingsKey
+        self.plexSSOHandler = PlexSSOHandler(
+            service: service,
+            settingsKey: settingsKey,
+            authContext: authContext
+        )
 
         loadSavedFilters()
-        if let savedToken = loadTokenFromKeychain() {
+        if let savedToken = plexSSOHandler.loadTokenFromKeychain() {
             sessionToken = savedToken
             Task {
                 do {
@@ -130,7 +135,7 @@ class OverseerrUsersViewModel: ObservableObject {
                     await OverseerrAuthManager.shared.probeSession()
                 } catch {
                     print("Failed to restore Plex token: \(error.localizedDescription)")
-                    deleteTokenFromKeychain()
+                    plexSSOHandler.deleteTokenFromKeychain()
                 }
             }
         }
@@ -631,27 +636,10 @@ class OverseerrUsersViewModel: ObservableObject {
         }
     }
 
-    private func saveTokenToKeychain(_ token: String) {
-        if let data = token.data(using: .utf8) {
-            KeychainHelper.save(key: tokenKey, data: data)
-        }
-    }
-
-    private func loadTokenFromKeychain() -> String? {
-        guard let data = KeychainHelper.load(key: tokenKey),
-              let token = String(data: data, encoding: .utf8)
-        else { return nil }
-        return token
-    }
-
-    private func deleteTokenFromKeychain() {
-        KeychainHelper.delete(key: tokenKey)
-        sessionToken = nil
-    }
-
     /// Log out the current user, clearing session cookies and stored token.
     func logout() {
-        deleteTokenFromKeychain()
+        plexSSOHandler.deleteTokenFromKeychain()
+        sessionToken = nil
         // Remove all cookies from shared storage
         let cookieStore = HTTPCookieStorage.shared
         cookieStore.cookies?.forEach { cookieStore.deleteCookie($0) }
@@ -664,54 +652,19 @@ class OverseerrUsersViewModel: ObservableObject {
         Task { await OverseerrAuthManager.shared.probeSession() }
     }
 
-    // MARK: – Plex SSO - Unchanged
-
-    private let clientID: String = {
-        let key = "PlexClientID"
-        if let existing = UserDefaults.standard.string(forKey: key) { return existing }
-        let uuid = UUID().uuidString
-        UserDefaults.standard.set(uuid, forKey: key)
-        return uuid
-    }()
-
-    private let productName = "Cantinarr"
+    // MARK: – Plex SSO
 
     func startPlexSSO(host _: String, port _: String?) {
         plexSSOTask?.cancel()
         plexSSOTask = Task {
             do {
-                let pin = try await fetchPlexPIN()
-                let plexURL = buildPlexOAuthURL(code: pin.code)
-                let web = ASWebAuthenticationSession(url: plexURL, callbackURLScheme: nil) { _, error in
-                    // This callback is expected to fire with a
-                    // `.canceledLogin` error once the user dismisses the
-                    // browser window. Ignore that specific case so the log
-                    // isn't polluted with a benign message.
-                    if let err = error as? ASWebAuthenticationSessionError,
-                       err.code != .canceledLogin
-                    {
-                        print("ASWebAuthenticationSession failed: \(err.localizedDescription)")
-                    }
-                }
-                web.presentationContextProvider = authContext
-                web.prefersEphemeralWebBrowserSession = true
-                web.start()
-
-                let authToken = try await pollForPlexToken(pinID: pin.id, code: pin.code)
-
-                try await service.loginWithPlexToken(authToken)
-                saveTokenToKeychain(authToken)
-                sessionToken = authToken
-
+                let token = try await plexSSOHandler.startLogin()
+                sessionToken = token
                 await OverseerrAuthManager.shared.probeSession()
                 plexSSOTask = nil
-
             } catch is CancellationError {
-                // Task was cancelled – ignore
                 plexSSOTask = nil
             } catch {
-                // If the user completed another login successfully before this
-                // task finished, don't revert the authenticated state.
                 if case .authenticated = OverseerrAuthManager.shared.value {
                     return
                 }
@@ -723,65 +676,6 @@ class OverseerrUsersViewModel: ObservableObject {
         }
     }
 
-    private struct PlexPIN: Decodable { let id: Int; let code: String; let authToken: String? }
-
-    private func fetchPlexPIN() async throws -> PlexPIN {
-        var req = URLRequest(url: URL(string: "https://plex.tv/api/v2/pins?strong=true")!)
-        req.httpMethod = "POST"
-        req.setValue("application/json", forHTTPHeaderField: "Accept")
-        req.setValue(productName, forHTTPHeaderField: "X-Plex-Product")
-        req.setValue(clientID, forHTTPHeaderField: "X-Plex-Client-Identifier")
-        let (data, resp) = try await URLSession.shared.data(for: req)
-        guard let http = resp as? HTTPURLResponse, http.statusCode == 201 else {
-            throw URLError(
-                .badServerResponse,
-                userInfo: [
-                    NSLocalizedDescriptionKey: "Failed to fetch Plex PIN (status: \((resp as? HTTPURLResponse)?.statusCode ?? 0))",
-                ]
-            )
-        }
-        return try JSONDecoder().decode(PlexPIN.self, from: data)
-    }
-
-    private func pollForPlexToken(pinID: Int, code _: String) async throws -> String {
-        for _ in 0 ..< 120 {
-            try await Task.sleep(nanoseconds: 1_000_000_000)
-            try Task.checkCancellation()
-            let url = URL(string: "https://plex.tv/api/v2/pins/\(pinID)")!
-            var req = URLRequest(url: url)
-            req.setValue("application/json", forHTTPHeaderField: "Accept")
-            req.setValue(clientID, forHTTPHeaderField: "X-Plex-Client-Identifier")
-            req.setValue(productName, forHTTPHeaderField: "X-Plex-Product")
-
-            let (data, resp) = try await URLSession.shared.data(for: req)
-            guard let http = resp as? HTTPURLResponse else { continue }
-            if http.statusCode == 404 { throw URLError(
-                .cancelled,
-                userInfo: [NSLocalizedDescriptionKey: "Plex PIN expired or invalid."]
-            ) }
-            guard http.statusCode == 200 else { continue }
-
-            guard let status = try? JSONDecoder().decode(PlexPIN.self, from: data) else { continue }
-            if let token = status.authToken, !token.isEmpty { return token }
-        }
-        throw URLError(.timedOut, userInfo: [NSLocalizedDescriptionKey: "Plex login timed out."])
-    }
-
-    private func buildPlexOAuthURL(code: String) -> URL {
-        var base = URLComponents()
-        base.scheme = "https"
-        base.host = "app.plex.tv"
-        base.path = "/auth"
-
-        var fragComps = URLComponents()
-        fragComps.queryItems = [
-            .init(name: "clientID", value: clientID),
-            .init(name: "code", value: code),
-            .init(name: "context[device][product]", value: productName),
-        ]
-        base.fragment = "?" + (fragComps.percentEncodedQuery ?? "")
-        return base.url!
-    }
 }
 
 extension OverseerrUsersViewModel: OverseerrPlexSSODelegate {
